Order Tracking State Machine + CreateNewOrderMutation (#25)
This PR implements Order Object for our DB, the CreateNewOrderMutation, and Order Resolvers. This consists of 3 key parts:

The Order Object model is as implemented in this spec on notion. This model will help us track an Order that a user makes as it moves through our system. This unlocks further work on the State Machine.
This implements the CreateNewOrderMutation and Order Resolvers. The CreateNewOrderMutation creates a new Order and CardVersion at the end of the card builder. The Card Version and Order are linked on the Order Object, and they will be linked to the User.
This also implements a few new mocks to help with our testing (createMockCard and createMockOrder).
This PR also moves the enumeration of Role and OrderState to the utils folder as we'll be using it throughout our system makes more sense to split them off